+++ src/test/java/com/caoccao/javet/values/reference/TestV8ValueObject.java
@@ -45,7 +45,7 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
                     new MockAnnotationBasedCallbackReceiver();
             List<JavetCallbackContext> javetCallbackContexts =
                     v8ValueObject.bind(mockAnnotationBasedCallbackReceiver);
-            assertEquals(20, javetCallbackContexts.size());
+            assertEquals(22, javetCallbackContexts.size());
             assertEquals(0, mockAnnotationBasedCallbackReceiver.getCount());
             assertEquals(123, v8Runtime.getExecutor("a.integerValue").executeInteger());
             assertEquals(1, mockAnnotationBasedCallbackReceiver.getCount());
@@ -63,9 +63,17 @@ public class TestV8ValueObject extends BaseTestJavetRuntime {
             assertEquals(9, mockAnnotationBasedCallbackReceiver.getCount());
             assertEquals("def", v8Runtime.getExecutor("a['stringValue']").executeString());
             assertEquals(10, mockAnnotationBasedCallbackReceiver.getCount());
-            assertEquals(18, v8ValueObject.unbind(mockAnnotationBasedCallbackReceiver));
+            assertNull(v8Runtime.getExecutor("a[Symbol.for('symbolValue')]").executeString());
+            assertEquals(11, mockAnnotationBasedCallbackReceiver.getCount());
+            v8Runtime.getExecutor("a[Symbol.for('symbolValue')] = 'abc';").executeVoid();
+            assertEquals(12, mockAnnotationBasedCallbackReceiver.getCount());
+            assertEquals("abc", mockAnnotationBasedCallbackReceiver.getSymbolValue());
+            assertEquals(13, mockAnnotationBasedCallbackReceiver.getCount());
+            assertEquals("abc", v8Runtime.getExecutor("a[Symbol.for('symbolValue')]").executeString());
+            assertEquals(14, mockAnnotationBasedCallbackReceiver.getCount());
+            assertEquals(19, v8ValueObject.unbind(mockAnnotationBasedCallbackReceiver));
             assertNull(v8Runtime.getExecutor("a['stringValue']").executeString());
-            assertEquals(10, mockAnnotationBasedCallbackReceiver.getCount());
+            assertEquals(14, mockAnnotationBasedCallbackReceiver.getCount());
         }
         v8Runtime.lowMemoryNotification();
     }