+++ src/test/java/com/caoccao/javet/values/reference/builtin/TestV8ValueBuiltInObject.java
@@ -18,7 +18,9 @@
 package com.caoccao.javet.values.reference.builtin;
 
 import com.caoccao.javet.BaseTestJavetRuntime;
+import com.caoccao.javet.entities.JavetEntitySymbol;
 import com.caoccao.javet.exceptions.JavetException;
+import com.caoccao.javet.values.reference.IV8ValueArray;
 import com.caoccao.javet.values.reference.V8ValueObject;
 import org.junit.jupiter.api.Test;
 
@@ -38,6 +40,19 @@ public class TestV8ValueBuiltInObject extends BaseTestJavetRuntime {
                 try (V8ValueObject v8ValueObjectC = V8ValueBuiltInObject.assign(v8ValueObjectA, v8ValueObjectB)) {
                     assertEquals("{\"x\":1,\"y\":2}", v8ValueObjectC.toJsonString());
                 }
+            }
+        }
+    }
+
+    @Test
+    public void testGetOwnPropertySymbols() throws JavetException {
+        try (V8ValueBuiltInObject v8ValueBuiltInObject = v8Runtime.getGlobalObject().getBuiltInObject();
+             V8ValueObject v8ValueObject = v8Runtime.getExecutor("const a = {};" +
+                     "a[Symbol.for('test')] = 1;" +
+                     "a;").execute()) {
+            try (IV8ValueArray v8ValueArray = v8ValueBuiltInObject.getOwnPropertySymbols(v8ValueObject)) {
+                assertEquals(1, v8ValueArray.getLength());
+                assertEquals("test", ((JavetEntitySymbol) v8ValueArray.getObject(0)).getDescription());
             }
         }
     }